+++ pkg/webui/account/views/forgot-password/forgot-password.styl
@@ -1,0 +1,29 @@
+// Copyright © 2020 The Things Network Foundation, The Things Industries B.V.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+.full-height
+  height: 100%
+  +media-query($bp.s)
+    height: auto
+
+.title
+  h1()
+  margin-bottom: $ls.xs
+  line-height: 1
+
+.description
+  h4()
+  font-weight: $fwh.normal
+  color: $tc-subtle-gray
+  margin-bottom: $ls.xs
