+++ pkg/webui/locales/en.json
@@ -1,4 +1,38 @@
 {
+  "account.views.authorize.index.modalTitle": "Request for permission",
+  "account.views.authorize.index.modalSubtitle": "{clientName} is requesting to be granted the following rights:",
+  "account.views.authorize.index.loginInfo": "You are logged in as {userId}.",
+  "account.views.authorize.index.redirectInfo": "You will be redirected to {redirectUri}",
+  "account.views.authorize.index.authorize": "Authorize",
+  "account.views.authorize.index.noDescription": "This client does not provide a description",
+  "account.views.code.index.code": "Your authorization code",
+  "account.views.create-account.index.createAccount": "Create a new {siteName} account",
+  "account.views.create-account.index.register": "Register",
+  "account.views.create-account.index.registrationApproved": "You have successfully registered and can login now",
+  "account.views.create-account.index.registrationPending": "You have successfully sent the registration request. Please wait until an admin approves it.",
+  "account.views.forgot-password.index.loginPage": "Login page",
+  "account.views.forgot-password.index.forgotPassword": "Forgot password",
+  "account.views.forgot-password.index.passwordRequested": "You will receive an email with reset instructions shortly",
+  "account.views.forgot-password.index.goToLogin": "Go to login",
+  "account.views.forgot-password.index.send": "Send",
+  "account.views.forgot-password.index.resetPasswordDescription": "Please enter your username to receive an email with reset instructions",
+  "account.views.forgot-password.index.requestTempPassword": "Reset password",
+  "account.views.login.index.createAccount": "Create an account",
+  "account.views.login.index.forgotPassword": "Forgot password?",
+  "account.views.login.index.loginToContinue": "Please login to continue",
+  "account.views.update-password.index.newPassword": "New password",
+  "account.views.update-password.index.oldPassword": "Old password",
+  "account.views.update-password.index.passwordChanged": "Password changed",
+  "account.views.update-password.index.revokeAccess": "Revoke access",
+  "account.views.update-password.index.logoutAllDevices": "Log out from all end devices",
+  "account.views.update-password.index.revokeWarning": "This will revoke access from all logged in devices",
+  "account.views.update-password.index.sessionRevoked": "Session revoked",
+  "account.views.validate.index.validateSuccess": "Contact info validated",
+  "account.views.validate.index.validateFail": "There was an error and the contact info could not be validated",
+  "account.views.validate.index.goToLogin": "Go to login",
+  "account.views.validate.index.tokenNotFoundTitle": "Token not found",
+  "account.views.validate.index.tokenNotFoundMessage": "The validation token was not found. This could mean that the contact info has already been validated. Otherwise, please contact an administrator.",
+  "account.views.validate.index.contactInfoValidation": "Contact info validation",
   "components.collapse.index.collapse": "Collapse",
   "components.collapse.index.expand": "Expand",
   "components.data-sheet.index.noData": "No data available",
@@ -730,39 +764,5 @@
   "lib.shared-messages.webhookBaseUrl": "Base URL",
   "lib.shared-messages.webhookFormat": "Webhook format",
   "lib.shared-messages.webhookId": "Webhook ID",
-  "lib.shared-messages.webhooks": "Webhooks",
-  "oauth.views.authorize.index.modalTitle": "Request for permission",
-  "oauth.views.authorize.index.modalSubtitle": "{clientName} is requesting to be granted the following rights:",
-  "oauth.views.authorize.index.loginInfo": "You are logged in as {userId}.",
-  "oauth.views.authorize.index.redirectInfo": "You will be redirected to {redirectUri}",
-  "oauth.views.authorize.index.authorize": "Authorize",
-  "oauth.views.authorize.index.noDescription": "This client does not provide a description",
-  "oauth.views.code.index.code": "Your authorization code",
-  "oauth.views.create-account.index.createAccount": "Create a new {siteName} account",
-  "oauth.views.create-account.index.register": "Register",
-  "oauth.views.create-account.index.registrationApproved": "You have successfully registered and can login now",
-  "oauth.views.create-account.index.registrationPending": "You have successfully sent the registration request. Please wait until an admin approves it.",
-  "oauth.views.forgot-password.index.loginPage": "Login page",
-  "oauth.views.forgot-password.index.forgotPassword": "Forgot password",
-  "oauth.views.forgot-password.index.passwordRequested": "You will receive an email with reset instructions shortly",
-  "oauth.views.forgot-password.index.goToLogin": "Go to login",
-  "oauth.views.forgot-password.index.send": "Send",
-  "oauth.views.forgot-password.index.resetPasswordDescription": "Please enter your username to receive an email with reset instructions",
-  "oauth.views.forgot-password.index.requestTempPassword": "Reset password",
-  "oauth.views.login.index.createAccount": "Create an account",
-  "oauth.views.login.index.forgotPassword": "Forgot password?",
-  "oauth.views.login.index.loginToContinue": "Please login to continue",
-  "oauth.views.update-password.index.newPassword": "New password",
-  "oauth.views.update-password.index.oldPassword": "Old password",
-  "oauth.views.update-password.index.passwordChanged": "Password changed",
-  "oauth.views.update-password.index.revokeAccess": "Revoke access",
-  "oauth.views.update-password.index.logoutAllDevices": "Log out from all end devices",
-  "oauth.views.update-password.index.revokeWarning": "This will revoke access from all logged in devices",
-  "oauth.views.update-password.index.sessionRevoked": "Session revoked",
-  "oauth.views.validate.index.validateSuccess": "Contact info validated",
-  "oauth.views.validate.index.validateFail": "There was an error and the contact info could not be validated",
-  "oauth.views.validate.index.goToLogin": "Go to login",
-  "oauth.views.validate.index.tokenNotFoundTitle": "Token not found",
-  "oauth.views.validate.index.tokenNotFoundMessage": "The validation token was not found. This could mean that the contact info has already been validated. Otherwise, please contact an administrator.",
-  "oauth.views.validate.index.contactInfoValidation": "Contact info validation"
+  "lib.shared-messages.webhooks": "Webhooks"
 }
