+++ pkg/webui/locales/xx.json
@@ -1,4 +1,38 @@
 {
+  "account.views.authorize.index.modalTitle": "Xxxxxxx xxx xxxxxxxxxx",
+  "account.views.authorize.index.modalSubtitle": "{clientName} xx xxxxxxxxxx xx xx xxxxxxx xxx xxxxxxxxx xxxxxx:",
+  "account.views.authorize.index.loginInfo": "Xxx xxx xxxxxx xx xx {userId}.",
+  "account.views.authorize.index.redirectInfo": "Xxx xxxx xx xxxxxxxxxx xx {redirectUri}",
+  "account.views.authorize.index.authorize": "Xxxxxxxxx",
+  "account.views.authorize.index.noDescription": "Xxxx xxxxxx xxxx xxx xxxxxxx x xxxxxxxxxxx",
+  "account.views.code.index.code": "Xxxx xxxxxxxxxxxxx xxxx",
+  "account.views.create-account.index.createAccount": "Xxxxxx x xxx {siteName} xxxxxxx",
+  "account.views.create-account.index.register": "Xxxxxxxx",
+  "account.views.create-account.index.registrationApproved": "Xxx xxxx xxxxxxxxxxxx xxxxxxxxxx xxx xxx xxxxx xxx",
+  "account.views.create-account.index.registrationPending": "Xxx xxxx xxxxxxxxxxxx xxxx xxx xxxxxxxxxxxx xxxxxxx. Xxxxxx xxxx xxxxx xx xxxxx xxxxxxxx xx.",
+  "account.views.forgot-password.index.loginPage": "Xxxxx xxxx",
+  "account.views.forgot-password.index.forgotPassword": "Xxxxxx xxxxxxxx",
+  "account.views.forgot-password.index.passwordRequested": "Xxx xxxx xxxxxxx xx xxxxx xxxx xxxxx xxxxxxxxxxxx xxxxxxx",
+  "account.views.forgot-password.index.goToLogin": "Xx xx xxxxx",
+  "account.views.forgot-password.index.send": "Xxxx",
+  "account.views.forgot-password.index.resetPasswordDescription": "Xxxxxx xxxxx xxxx xxxxxxxx xx xxxxxxx xx xxxxx xxxx xxxxx xxxxxxxxxxxx",
+  "account.views.forgot-password.index.requestTempPassword": "Xxxxx xxxxxxxx",
+  "account.views.login.index.createAccount": "Xxxxxx xx xxxxxxx",
+  "account.views.login.index.forgotPassword": "Xxxxxx xxxxxxxx?",
+  "account.views.login.index.loginToContinue": "Xxxxxx xxxxx xx xxxxxxxx",
+  "account.views.update-password.index.newPassword": "Xxx xxxxxxxx",
+  "account.views.update-password.index.oldPassword": "Xxx xxxxxxxx",
+  "account.views.update-password.index.passwordChanged": "Xxxxxxxx xxxxxxx",
+  "account.views.update-password.index.revokeAccess": "Xxxxxx xxxxxx",
+  "account.views.update-password.index.logoutAllDevices": "Xxx xxx xxxx xxx xxx xxxxxxx",
+  "account.views.update-password.index.revokeWarning": "Xxxx xxxx xxxxxx xxxxxx xxxx xxx xxxxxx xx xxxxxxx",
+  "account.views.update-password.index.sessionRevoked": "Xxxxxxx xxxxxxx",
+  "account.views.validate.index.validateSuccess": "Xxxxxxx xxxx xxxxxxxxx",
+  "account.views.validate.index.validateFail": "Xxxxx xxx xx xxxxx xxx xxx xxxxxxx xxxx xxxxx xxx xx xxxxxxxxx",
+  "account.views.validate.index.goToLogin": "Xx xx xxxxx",
+  "account.views.validate.index.tokenNotFoundTitle": "Xxxxx xxx xxxxx",
+  "account.views.validate.index.tokenNotFoundMessage": "Xxx xxxxxxxxxx xxxxx xxx xxx xxxxx. Xxxx xxxxx xxxx xxxx xxx xxxxxxx xxxx xxx xxxxxxx xxxx xxxxxxxxx. Xxxxxxxxx, xxxxxx xxxxxxx xx xxxxxxxxxxxxx.",
+  "account.views.validate.index.contactInfoValidation": "Xxxxxxx xxxx xxxxxxxxxx",
   "components.collapse.index.collapse": "Xxxxxxxx",
   "components.collapse.index.expand": "Xxxxxx",
   "components.data-sheet.index.noData": "Xx xxxx xxxxxxxxx",
@@ -730,39 +764,5 @@
   "lib.shared-messages.webhookBaseUrl": "Xxxx XXX",
   "lib.shared-messages.webhookFormat": "Xxxxxxx xxxxxx",
   "lib.shared-messages.webhookId": "Xxxxxxx XX",
-  "lib.shared-messages.webhooks": "Xxxxxxxx",
-  "oauth.views.authorize.index.modalTitle": "Xxxxxxx xxx xxxxxxxxxx",
-  "oauth.views.authorize.index.modalSubtitle": "{clientName} xx xxxxxxxxxx xx xx xxxxxxx xxx xxxxxxxxx xxxxxx:",
-  "oauth.views.authorize.index.loginInfo": "Xxx xxx xxxxxx xx xx {userId}.",
-  "oauth.views.authorize.index.redirectInfo": "Xxx xxxx xx xxxxxxxxxx xx {redirectUri}",
-  "oauth.views.authorize.index.authorize": "Xxxxxxxxx",
-  "oauth.views.authorize.index.noDescription": "Xxxx xxxxxx xxxx xxx xxxxxxx x xxxxxxxxxxx",
-  "oauth.views.code.index.code": "Xxxx xxxxxxxxxxxxx xxxx",
-  "oauth.views.create-account.index.createAccount": "Xxxxxx x xxx {siteName} xxxxxxx",
-  "oauth.views.create-account.index.register": "Xxxxxxxx",
-  "oauth.views.create-account.index.registrationApproved": "Xxx xxxx xxxxxxxxxxxx xxxxxxxxxx xxx xxx xxxxx xxx",
-  "oauth.views.create-account.index.registrationPending": "Xxx xxxx xxxxxxxxxxxx xxxx xxx xxxxxxxxxxxx xxxxxxx. Xxxxxx xxxx xxxxx xx xxxxx xxxxxxxx xx.",
-  "oauth.views.forgot-password.index.loginPage": "Xxxxx xxxx",
-  "oauth.views.forgot-password.index.forgotPassword": "Xxxxxx xxxxxxxx",
-  "oauth.views.forgot-password.index.passwordRequested": "Xxx xxxx xxxxxxx xx xxxxx xxxx xxxxx xxxxxxxxxxxx xxxxxxx",
-  "oauth.views.forgot-password.index.goToLogin": "Xx xx xxxxx",
-  "oauth.views.forgot-password.index.send": "Xxxx",
-  "oauth.views.forgot-password.index.resetPasswordDescription": "Xxxxxx xxxxx xxxx xxxxxxxx xx xxxxxxx xx xxxxx xxxx xxxxx xxxxxxxxxxxx",
-  "oauth.views.forgot-password.index.requestTempPassword": "Xxxxx xxxxxxxx",
-  "oauth.views.login.index.createAccount": "Xxxxxx xx xxxxxxx",
-  "oauth.views.login.index.forgotPassword": "Xxxxxx xxxxxxxx?",
-  "oauth.views.login.index.loginToContinue": "Xxxxxx xxxxx xx xxxxxxxx",
-  "oauth.views.update-password.index.newPassword": "Xxx xxxxxxxx",
-  "oauth.views.update-password.index.oldPassword": "Xxx xxxxxxxx",
-  "oauth.views.update-password.index.passwordChanged": "Xxxxxxxx xxxxxxx",
-  "oauth.views.update-password.index.revokeAccess": "Xxxxxx xxxxxx",
-  "oauth.views.update-password.index.logoutAllDevices": "Xxx xxx xxxx xxx xxx xxxxxxx",
-  "oauth.views.update-password.index.revokeWarning": "Xxxx xxxx xxxxxx xxxxxx xxxx xxx xxxxxx xx xxxxxxx",
-  "oauth.views.update-password.index.sessionRevoked": "Xxxxxxx xxxxxxx",
-  "oauth.views.validate.index.validateSuccess": "Xxxxxxx xxxx xxxxxxxxx",
-  "oauth.views.validate.index.validateFail": "Xxxxx xxx xx xxxxx xxx xxx xxxxxxx xxxx xxxxx xxx xx xxxxxxxxx",
-  "oauth.views.validate.index.goToLogin": "Xx xx xxxxx",
-  "oauth.views.validate.index.tokenNotFoundTitle": "Xxxxx xxx xxxxx",
-  "oauth.views.validate.index.tokenNotFoundMessage": "Xxx xxxxxxxxxx xxxxx xxx xxx xxxxx. Xxxx xxxxx xxxx xxxx xxx xxxxxxx xxxx xxx xxxxxxx xxxx xxxxxxxxx. Xxxxxxxxx, xxxxxx xxxxxxx xx xxxxxxxxxxxxx.",
-  "oauth.views.validate.index.contactInfoValidation": "Xxxxxxx xxxx xxxxxxxxxx"
+  "lib.shared-messages.webhooks": "Xxxxxxxx"
 }
