--- pkg/webui/oauth/store/actions/init.js
@@ -1,23 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-export const INITIALIZE = 'INITIALIZE_REQUEST'
-export const INITIALIZE_SUCCESS = 'INITIALIZE_SUCCESS'
-export const INITIALIZE_FAILURE = 'INITIALIZE_FAILURE'
-
-export const initialize = () => ({ type: INITIALIZE })
-
-export const initializeSuccess = () => ({ type: INITIALIZE_SUCCESS })
-
-export const initializeFailure = error => ({ type: INITIALIZE_FAILURE, error })
