--- pkg/webui/oauth/views/authorize/authorize.styl
@@ -1,55 +1,0 @@
-// Copyright © 2019 The Things Network Foundation, The Things Industries B.V.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-.left
-  flex-basis: 65%
-  padding-right: $cs.m
-  +media-query($bp.s)
-    flex-basis: 100%
-    flex-grow: 0
-    margin-bottom: $ls.m
-
-  ul
-    text-margin-top()
-    padding-left: 0
-
-  li
-    display: flex
-    align-items: flex-start
-
-  li > span:first-child
-    color: $c-info
-    margin-right: $cs.xs
-    nudge('down', 3px)
-
-.right
-  +media-query($bp.s)
-    flex-grow: 0
-    padding-top: 3rem
-    background: url('../../../assets/misc/oauth-arrow-hor.svg') no-repeat
-  +media-query-min($bp.s)
-    flex-basis: 35%
-    padding-left: 3rem
-    background: url('../../../assets/misc/oauth-arrow.svg') no-repeat
-
-  h3
-    one-liner()
-    margin-top: 0
-    margin-bottom: $cs.s
-
-.login-info
-  color: $tc-deep-gray
-
-.no-description
-  color: $tc-subtle-gray
